requirements.txt
# Flask et extensions
Flask==3.0.2
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
Flask-Mail==0.9.1

# Base de données
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
alembic==1.13.1

# Gestion des fichiers et stockage
cloudinary==1.38.0
Werkzeug==3.0.1
python-magic==0.4.27
Pillow==10.2.0

# Traitement de documents
python-docx==1.1.0
mammoth==1.7.1
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.10
opencv-python==4.9.0.80
fitz==0.0.1.dev2
PyMuPDF==1.23.22
python-magic-bin==0.4.14

# Sécurité
PyJWT==2.8.0
bcrypt==4.1.2
cryptography==42.0.2

# Utilitaires
python-dateutil==2.8.2
requests==2.31.0
pytz==2024.1
six==1.16.0

# Logging et debugging
logging==0.4.9.6
pytest==8.0.0
black==24.1.1
flake8==7.0.0